addFoldersReducer
