+++ src/reducers/filesReducer.js
@@ -1,12 +1,17 @@
-import { FETCH_FILES_SUCCESS } from "../actions/filesActions"
+import { FETCH_FILES_SUCCESS, ADD_FILE_ACTION } from "../actions/filesActions"
 
 const initialState = []
 
-const files = (prevState = initialState, action) => {
+export const files = (prevState = initialState, action) => {
   if (action.type === FETCH_FILES_SUCCESS) {
     return action.response
   }
   return prevState
 }
 
-export default files
+export const addFiles = (prevState = initialState, action) => {
+  if (action.type === ADD_FILE_ACTION) {
+    return [...prevState, action.response]
+  }
+  return prevState
+}
