+++ src/reducers/foldersReducer.js
@@ -1,58 +1,18 @@
 import { FETCH_FOLDERS_SUCCESS } from "../actions/foldersActions"
+import { ADD_FOLDERS_SUCCCESS } from "./action/foldersActions"
 
-const initialState = [
-  {
-    _id: "dir_DOl2kN3n9lMeedN90kL9",
-    object: "directory",
-    name: "SammTrading",
-    created: "2018-03-29T00:00:00+00:00",
-    modified: "2018-03-29T00:00:00+00:00",
-    shares: [],
-    files: [],
-    dirs: [
-      {
-        _id: "dir_EoeInRgUgzMPh0aLN2nz",
-        object: "directory",
-        name: "Webshop",
-        created: "2018-03-29T11:36:42+00:00",
-        modified: "2018-03-29T11:36:42+00:00",
-        shares: []
-      }
-    ],
-    path: [
-      {
-        _id: "dir_kL9D9lMe3ndN90eOl2kN",
-        object: "directory",
-        name: "Meduza",
-        created: "2018-03-29T00:00:00+00:00",
-        modified: "2018-03-29T00:00:00+00:00",
-        index: 0
-      },
-      {
-        _id: "dir_3ndN90kL9D9lMeeOl2kN",
-        object: "directory",
-        name: "Clients",
-        created: "2018-03-29T00:00:00+00:00",
-        modified: "2018-03-29T00:00:00+00:00",
-        index: 1
-      }
-    ],
-    parent: {
-      _id: "dir_3ndN90kL9D9lMeeOl2kN",
-      object: "directory",
-      name: "Clients",
-      created: "2018-03-29T00:00:00+00:00",
-      modified: "2018-03-29T00:00:00+00:00",
-      index: 1
-    }
-  }
-]
+const initialState = []
 
-const folders = (prevState = initialState, action) => {
+export const folders = (prevState = initialState, action) => {
   if (action.type === FETCH_FOLDERS_SUCCESS) {
     return action.response
   }
   return prevState
 }
 
-export default folders
+export const addFolders = (prevState = initialState, action) => {
+  if (action.type === ADD_FOLDERS_SUCCCESS) {
+    return action.response
+  }
+  return prevState
+}
